Revert move of TopLeftCopToMultiple back to johnson
